+++ library/Gc/Mvc/Service/ViewTemplatePathStackFactory.php
@@ -1,0 +1,75 @@
+<?php
+/**
+ * This source file is part of GotCms.
+ *
+ * GotCms is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Lesser General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * GotCms is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License along
+ * with GotCms. If not, see <http://www.gnu.org/licenses/lgpl-3.0.html>.
+ *
+ * PHP Version >=5.3
+ *
+ * @category   Gc
+ * @package    Library
+ * @subpackage Mvc\Service
+ * @author     Pierre Rambaud (GoT) <pierre.rambaud86@gmail.com>
+ * @license    GNU/LGPL http://www.gnu.org/licenses/lgpl-3.0.html
+ * @link       http://www.got-cms.com
+ */
+
+namespace Gc\Mvc\Service;
+
+use Zend\ServiceManager\FactoryInterface;
+use Zend\ServiceManager\ServiceLocatorInterface;
+use Gc\View\Resolver\TemplatePathStack;
+
+/**
+ * Create modules via service
+ *
+ * @category   Gc
+ * @package    Library
+ * @subpackage Mvc\Service
+ */
+class ViewTemplatePathStackFactory implements FactoryInterface
+{
+    /**
+     * Create the template map view resolver
+     *
+     * Creates a Zend\View\Resolver\TemplatePathStack and populates it with the
+     * ['view_manager']['template_path_stack'] and sets the default suffix with the
+     * ['view_manager']['default_template_suffix']
+     *
+     * @param ServiceLocatorInterface $serviceLocator Service Locator
+     *
+     * @return ViewResolver\TemplatePathStack
+     */
+    public function createService(ServiceLocatorInterface $serviceLocator)
+    {
+        $config = $serviceLocator->get('Config');
+
+        $templatePathStack = new TemplatePathStack();
+        $templatePathStack->setUseStreamWrapper(true);
+
+        if (is_array($config) && isset($config['view_manager'])) {
+            $config = $config['view_manager'];
+            if (is_array($config)) {
+                if (isset($config['template_path_stack'])) {
+                    $templatePathStack->addPaths($config['template_path_stack']);
+                }
+                if (isset($config['default_template_suffix'])) {
+                    $templatePathStack->setDefaultSuffix($config['default_template_suffix']);
+                }
+            }
+        }
+
+        return $templatePathStack;
+    }
+}